+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   causing count(*) queries on large time series to use
   lots of memory
 - [Issue #437](https://github.com/influxdb/influxdb/issues/437). Queries with negative constants don't parse properly
+- [Issue #432](https://github.com/influxdb/influxdb/issues/432). Deleted data using a delete query is resurrected after a server restart
 
 ## v0.5.6 [2014-04-08]